+++ website/otfw.config.js
@@ -1,0 +1,5 @@
+export default {
+  site: {
+    url: "https://micro-ui.opentechf.org",
+  },
+};